product CRUDL

+++ app/delete_product.go
@@ -1,8 +1,10 @@
 package app
 
 import (
+	"encoding/json"
 	"net/http"
 
+	service "bitbucket.org/andyfusniakteam/ecom-api-go/service/firebase"
 	"github.com/go-chi/chi"
 	log "github.com/sirupsen/logrus"
 )
@@ -16,6 +18,19 @@ func (a *App) DeleteProductHandler() http.HandlerFunc {
 
 		productID := chi.URLParam(r, "id")
 		if err := a.Service.DeleteProduct(ctx, productID); err != nil {
+			if err == service.ErrProductNotFound {
+				w.WriteHeader(http.StatusNotFound) // 404 Not Found
+				json.NewEncoder(w).Encode(struct {
+					Status  int    `json:"status"`
+					Code    string `json:"code"`
+					Message string `json:"message"`
+				}{
+					http.StatusNotFound,
+					ErrCodeProductNotFound,
+					"product not found",
+				})
+				return
+			}
 			contextLogger.Errorf("a.Service.DeleteProduct(ctx, productID=%q) failed: %v", productID, err)
 			w.WriteHeader(http.StatusInternalServerError) // 500 Internal Server Error
 			return
